Update README.md
Along with stock and equity markets, there is a new investment market dominated by new currency known 
as cryptocurrency. Cryptocurrency is the new age digital money which has impacted the financial world, since 
the launch of its first crypto-coin called Bitcoin Nakamoto (2008). Investment companies ranging from public 
sector banks to private investments have been encouraging their client profiles to invest their funds into these 
cryptocurrencies for sooner, faster and bigger returns. The reason cryptocurrencies are getting popular is due 
to their various advantages ranging from being fast, secure, scalable, trustworthy, reliable, decentralized and 
more
Cryptocurrency analysis and predictions have been sought after due to the increasing values of these cryptocoins since their inception. Cryptocurrency prediction and price analysis has emerged one of the popular 
trends in the prognostication domain. Prediction of future prices of cryptocurrencies using past data is 
possible with the help of statistical tools and techniques. The aim of this research is divided into two parts. 
The first part is to determine to what magnitude (positive/negative) are the different cryptocurrency prices 
(Bitcoin, Ethereum, Binance coin, Tether, and Ripple) correlated with each other, measured using statistical 
techniques Pearson correlation coefficients and To what extent can the prices of the different 
cryptocurrencies (Bitcoin, Ethereum, Binance coin, Tether, and Ripple) be predicted using Facebook prophet
since the significance of forecasting can be deemed one of the essential factors when gaining an edge in any 
kind of investment. The second part of the research deals with the risk and returns associated with these top 
5 currencies since the skyrocketing price rise of cryptocurrencies, investors looking forward to a right 
opportunity to invest in this new age digital money which is now a commodity.


Cryptocurrency, generally referred to as cryptocurrency or crypto, may be any sort of currency that exists 
digitally or nearly and uses cryptography to secure transactions. Cryptocurrencies don't have a central 
issuance or regulation authority, rather than employing a suburbanized system to record transactions and 
issue new units.
1. Bitcoin (BTC) - Created in 2009 by someone under the pseudonym Satoshi Nakamoto, Bitcoin (BTC) is 
the original cryptocurrency. As with most cryptocurrencies, BTC runs on a blockchain, or a ledger logging 
transaction distributed across a network of thousands of computers. Because additions to the 
distributed ledgers must be verified by solving a cryptographic puzzle, a process called proof of work, 
Bitcoin is kept secure and safe from fraudsters.
2. Ethereum (ETH) - Both a cryptocurrency and a blockchain platform, Ethereum is a favourite of program 
developers because of its potential applications, like so-called smart contracts that automatically 
execute when conditions are met and non-fungible tokens (NFTs).
3. Binance Coin(BNB) - The Binance Coin is a form of cryptocurrency that you can use to trade and pay 
fees on Binance, one of the largest crypto exchanges in the world.
4. Tether(USDT) - Unlike some other forms of cryptocurrency, Tether is a stable coin, meaning it’s backed 
by fiat currencies like U.S. dollars and the Euro and hypothetically keeps a value equal to one of those 
denominations. In theory, this means Tether’s value is supposed to be more consistent than other 
cryptocurrencies, and it’s favoured by investors who are wary of the extreme volatility of other coins
5. Ripple(XRP) - Created by Ripple Labs Inc., a US-based technology company. Released in 2012, a digital 
technology and payment processing company, XRP can be used on that network to facilitate exchanges 
of different currency types, including fiat currencies and other major cryptocurrencies.

This study is based on secondary data which has been collected from the website 
https://www.kaggle.com/nshreya788/eda-on-top-5-cryptocurrencies The dataset contains 5 sheets namely as BTC, 
ETH, BNB, USDT, XRP. The data attributes are as follows:
1. Date: Date of observation
2. Close: Price on the given day in USD(Also the closing price for that day)
3. Open: Opening price on the given day
4. High: Highest price on the given day
5. Low: Lowest price on the given day
6. Volume: Volume of transactions on the given day 
The data from the last five years (from the 9th November, 2017 to 31st December 2021) were included in the study.

Primary Data Source: https://finance.yahoo.com/cryptocurrencies/